Fix: Implementacion de documentacion.

+++ backend_projects/prod/db_models/project_db_model.py
@@ -64,7 +64,11 @@ class ProjectDBModel(db.Model):
 
     @staticmethod
     def delete(deleted_id):
+        projects_query = ProjectDBModel.query.filter_by(id=deleted_id)
+        if projects_query.count() == 0:
+            return 1
         deleted_objects = ProjectDBModel.__table__.delete().where(
             ProjectDBModel.id == deleted_id)
         db.session.execute(deleted_objects)
         db.session.commit()
+        return 0
